requirements.txt
asgiref==3.7.2
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.3
djangorestframework==3.14.0
httpie==3.2.2
idna==3.6
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
numba==0.59.0
numpy==1.26.4
packaging==23.2
platformdirs==4.2.0
pooch==1.8.1
pycparser==2.21
Pygments==2.17.2
PySocks==1.7.1
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.1.1
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1